Create README.md
well.. i've been trying to be a part of the human solution to the human problem for all of my life.  i've just run a hopefully updated "Hijack,this."  I  feel it is a genuine work of devotion.  but i am unable to know how to interpret the  report, as regards categories  e.g.    a)  "friendly"  and  b} "unfriendly" .. so to speak...  & some clear.. "Simple"  guidance   would be welcome.. Thanks.  ianbe .. or (maybe)  "@ian374" .. etc..